+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
-## [Unreleased]
+## [0.39.0]
 
+### Added
+
+- `CueEnterEvent.image` and `CueExitEvent.image` to expose the Base64 encoded image data URI of the cue when available
 
 ## [0.38.0] - 2025-02-28